LayoutTests/imported/w3c/web-platform-tests/cors/credentials-flag-expected.txt

Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-cookie.py?ident=1489438094875_1
CONSOLE MESSAGE: line 40: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-cookie.py?ident=1489438094875_1 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-cookie.py?ident=1489438094877_2
CONSOLE MESSAGE: line 70: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-cookie.py?ident=1489438094877_2 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-cookie.py?ident=1489438094879_3
CONSOLE MESSAGE: line 100: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-cookie.py?ident=1489438094879_3 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=TRUE
CONSOLE MESSAGE: line 118: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=TRUE due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=True
CONSOLE MESSAGE: line 118: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=True due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=%22true%22
CONSOLE MESSAGE: line 118: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=%22true%22 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=false
CONSOLE MESSAGE: line 118: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=false due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=1
CONSOLE MESSAGE: line 118: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=1 due to access control checks.
Blocked access to external URL http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=0
CONSOLE MESSAGE: line 118: XMLHttpRequest cannot load http://www1.localhost:8800/cors/resources/cors-makeheader.py?credentials=0 due to access control checks.
CORS - Access-Control-Allow-Credentials


Harness Error (TIMEOUT), message = null

PASS Setting withCredentials on a sync XHR object should not throw 
TIMEOUT Don't send cookie by default Test timed out
TIMEOUT Don't send cookie part 2 Test timed out
TIMEOUT Don't obey Set-Cookie when withCredentials=false Test timed out
PASS Access-Control-Allow-Credentials: TRUE should be disallowed (async) 
PASS Access-Control-Allow-Credentials: True should be disallowed (async) 
PASS Access-Control-Allow-Credentials: "true" should be disallowed (async) 
PASS Access-Control-Allow-Credentials: false should be disallowed (async) 
PASS Access-Control-Allow-Credentials: 1 should be disallowed (async) 
PASS Access-Control-Allow-Credentials: 0 should be disallowed (async) 
